agent/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
fastapi==0.124.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==1.3.0
langchain-core==1.2.0
langchain-openai==1.1.2
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
numpy==2.3.5
openai==2.11.0
orjson==3.11.5
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pypandoc==1.16.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
uuid_utils==0.12.0
xxhash==3.6.0
zstandard==0.25.0
